Make source dependencies configurable.
Currently the source dependencies are not configurable which makes it
impossible to switch dependencies in some cases.  For example, I needed
to compile PHP with curl using libcurl4-openssl-dev but this recipe
hard encodes the use of libcurl4-gnutls-dev.

Using a wrapper recipe does not help in this case because both of the
packages save files that are named exactly the same thing to exactly the
same location (the second install always overrides the first).

This change ensures that the user has override control over these
dependencies if so desired.