Switch from maintainers files to a simple readme section
We weren't actually updating the maintainers files or doing anything with the TOML files. I like the idea, but it ended up being more work than it's *currently* worth (emphasis on the current part). This switches us to a simpler blurb in the readme the links to our centrally located docs. We can more easily update those docs in Github instead of updating 100+ repos. If we bot the world in the future we can / should bring back the toml setup (except yaml because it's cooler)

Signed-off-by: Tim Smith <tsmith@chef.io>